checkpoint 2: refactor: component based
